+++ Core/GPUAlgorithms/Interactions/sort_particles.jl
@@ -16,7 +16,7 @@ function sort_particles!(cells_data, particles, num_particles)
         cells_data.cell_starts_scratch,
         cells_data.cell_indices,
         particles,
-        Int32(num_particles);
+        Int64(num_particles);
         ndrange=total_num_threads)
     # KernelAbstractions.synchronize(CUDABackend())
 end #function
@@ -29,12 +29,12 @@ end #function
     @Const(particles),
     num_particles)
 
-    I = Int32(@index(Global, Linear))
-    stride = Int32(@ndrange()[1])
+    I = Int64(@index(Global, Linear))
+    stride = Int64(@ndrange()[1])
 
     for i = I:stride:num_particles
         c = cell_indices[i]
-        pos = CUDA.atomic_add!(pointer(cell_starts_scratch, c), Int32(1))
+        pos = CUDA.atomic_add!(pointer(cell_starts_scratch, c), Int64(1))
         sorted_particles[pos] = particles[i]
         perm[pos] = i
     end #for i
